Moved client model distributed stuff into proper package
